+++ app.py
@@ -3,80 +3,36 @@
 
 from flask import Flask, render_template, request, jsonify, send_from_directory, redirect, url_for, flash, Response, abort
 import os
-import json
 import time
 import subprocess
 import threading
 import asyncio
 import requests
-import cv2
-import shutil
+import logging
 import signal
 import atexit
 import base64
 import sys
 from datetime import datetime
-from werkzeug.utils import secure_filename
-from telegram import Bot
-from telegram.error import TelegramError
 from runware import Runware, IImageInference
-from PIL import Image
+from config_utils import (
+    PHOTOS_FOLDER,
+    EFFECT_FOLDER,
+    load_config,
+    save_config,
+    ensure_directories,
+)
+from camera_utils import UsbCamera, detect_cameras
+from telegram_utils import send_to_telegram
 
 app = Flask(__name__)
-app.secret_key = 'photobooth_secret_key_2024'
+app.secret_key = os.environ.get('SECRET_KEY', 'photobooth_secret_key_2024')
 
-# Configuration
-PHOTOS_FOLDER = 'photos'
-EFFECT_FOLDER = 'effet'
-CONFIG_FILE = 'config.json'
-ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg'}
+logging.basicConfig(level=logging.INFO, format='[%(levelname)s] %(message)s')
+logger = logging.getLogger(__name__)
 
-# Créer les dossiers s'ils n'existent pas
-print(f"[DEBUG] Création du dossier photos: {PHOTOS_FOLDER}")
-os.makedirs(PHOTOS_FOLDER, exist_ok=True)
-print(f"[DEBUG] Création du dossier effet: {EFFECT_FOLDER}")
-os.makedirs(EFFECT_FOLDER, exist_ok=True)
-print(f"[DEBUG] Dossiers créés - Photos: {os.path.exists(PHOTOS_FOLDER)}, Effet: {os.path.exists(EFFECT_FOLDER)}")
-
-# Configuration par défaut
-DEFAULT_CONFIG = {
-    'footer_text': 'Photobooth',
-    'timer_seconds': 3,
-    'high_density': False,
-    'slideshow_enabled': False,
-    'slideshow_delay': 60,
-    'slideshow_source': 'photos',
-    'effect_enabled': False,
-    'effect_prompt': 'Transform this photo into a beautiful ghibli style',
-    'effect_steps': 5,
-    'runware_api_key': '',
-    'telegram_enabled': False,
-    'telegram_bot_token': '',
-    'telegram_chat_id': '',
-    'telegram_send_type': 'photos',
-    'camera_type': 'picamera',  # 'picamera' ou 'usb'
-    'usb_camera_id': 0,  # ID de la caméra USB (généralement 0 pour la première caméra)
-    # Paramètres d'imprimante
-    'printer_enabled': True,
-    'printer_port': '/dev/AMA0',  # Port série de l'imprimante
-    'printer_baudrate': 9600,  # Vitesse de communication
-    'print_resolution': 384  # Résolution d'impression
-}
-
-def load_config():
-    """Charger la configuration depuis le fichier JSON"""
-    if os.path.exists(CONFIG_FILE):
-        try:
-            with open(CONFIG_FILE, 'r', encoding='utf-8') as f:
-                return json.load(f)
-        except:
-            pass
-    return DEFAULT_CONFIG.copy()
-
-def save_config(config_data):
-    """Sauvegarder la configuration dans un fichier JSON"""
-    with open(CONFIG_FILE, 'w', encoding='utf-8') as f:
-        json.dump(config_data, f, indent=2, ensure_ascii=False)
+# Initialiser les dossiers nécessaires
+ensure_directories()
 
 def check_printer_status():
     """Vérifier l'état de l'imprimante thermique"""
@@ -139,71 +95,6 @@ def check_printer_status():
             'paper_status': 'unknown'
         }
 
-async def _send_telegram_photo(bot_token, chat_id, photo_path, caption):
-    """Fonction asynchrone pour envoyer une photo via Telegram"""
-    bot = Bot(token=bot_token)
-    
-    # Vérifier et nettoyer l'ID du chat
-    cleaned_chat_id = chat_id.strip()
-    
-    # Si c'est un nom d'utilisateur ou canal, s'assurer qu'il commence par @
-    if cleaned_chat_id and cleaned_chat_id[0].isalpha() and not cleaned_chat_id.startswith('@'):
-        cleaned_chat_id = '@' + cleaned_chat_id
-    
-    print(f"[TELEGRAM] Utilisation de l'ID de chat: '{cleaned_chat_id}'")
-    
-    try:
-        with open(photo_path, 'rb') as photo_file:
-            await bot.send_photo(
-                chat_id=cleaned_chat_id,
-                photo=photo_file,
-                caption=caption
-            )
-    except Exception as e:
-        if "chat not found" in str(e).lower():
-            print(f"[TELEGRAM] ERREUR: Chat introuvable avec l'ID '{cleaned_chat_id}'")
-            print("[TELEGRAM] Assurez-vous que:")
-            print("   - Le bot a été ajouté au groupe/canal")
-            print("   - Pour un groupe: l'ID commence par '-' (ex: -123456789)")
-            print("   - Pour un canal: utilisez '@nom_du_canal' ou ajoutez le bot comme admin")
-            print("   - Pour un chat privé: utilisez l'ID numérique de l'utilisateur")
-        raise
-
-def send_to_telegram(photo_path, photo_type="photo"):
-    """Envoyer une photo sur Telegram"""
-    if not config.get('telegram_enabled', False):
-        return
-    
-    bot_token = config.get('telegram_bot_token', '')
-    chat_id = config.get('telegram_chat_id', '')
-    
-    if not bot_token or not chat_id:
-        print("[TELEGRAM] Configuration incomplète (token ou chat_id manquant)")
-        return
-    
-    try:
-        print(f"[TELEGRAM] Envoi de {photo_path} vers le chat {chat_id}")
-        
-        # Préparer le message
-        caption = f"📸 Nouvelle photo du photobooth!"
-        if photo_type == "effet":
-            caption = f"🎨 Photo avec effet IA du photobooth!"
-        
-        # Exécuter la coroutine dans une nouvelle boucle asyncio
-        async def send_photo_async():
-            try:
-                await _send_telegram_photo(bot_token, chat_id, photo_path, caption)
-                print("[TELEGRAM] Photo envoyée avec succès!")
-            except Exception as e:
-                print(f"[TELEGRAM] Erreur dans la coroutine: {e}")
-                
-        # Exécuter dans une nouvelle boucle d'événements asyncio
-        asyncio.run(send_photo_async())
-        
-    except TelegramError as e:
-        print(f"[TELEGRAM] Erreur Telegram: {e}")
-    except Exception as e:
-        print(f"[TELEGRAM] Erreur lors de l'envoi: {e}")
 
 # Fonction pour détecter les ports série disponibles
 def detect_serial_ports():
@@ -241,254 +132,10 @@ def detect_serial_ports():
         if sys.platform.startswith('win'):
             available_ports = [('COM1', 'COM1'), ('COM3', 'COM3')]
         else:
-            available_ports = [('/dev/ttyAMA0', '/dev/AMA0'), ('/dev/ttyS0', '/dev/ttyS0')]
+            available_ports = [('/dev/ttyAMA0', '/dev/ttyAMA0'), ('/dev/ttyS0', '/dev/ttyS0')]
     
     return available_ports
 
-# Fonction pour détecter les caméras USB disponibles
-def detect_cameras():
-    """Détecter les caméras USB disponibles et retourner une liste de (id, nom)"""
-    available_cameras = []
-    print("[CAMERA] Début de la détection des caméras USB...")
-    
-    # Tester les 10 premiers indices de caméra (0-9) pour être plus exhaustif
-    for i in range(10):
-        try:
-            print(f"[CAMERA] Test de la caméra ID {i}...")
-            
-            # Essayer différents backends OpenCV
-            backends = [cv2.CAP_ANY, cv2.CAP_DSHOW, cv2.CAP_V4L2, cv2.CAP_GSTREAMER]
-            cap = None
-            
-            for backend in backends:
-                try:
-                    cap = cv2.VideoCapture(i, backend)
-                    if cap.isOpened():
-                        # Configurer une résolution plus élevée pour tester les capacités
-                        resolutions_to_test = [
-                            (1920, 1080),  # Full HD
-                            (1280, 720),   # HD
-                            (640, 480)     # VGA (fallback)
-                        ]
-                        
-                        best_resolution = None
-                        best_fps = 0
-                        
-                        for test_width, test_height in resolutions_to_test:
-                            # Essayer de configurer cette résolution
-                            cap.set(cv2.CAP_PROP_FRAME_WIDTH, test_width)
-                            cap.set(cv2.CAP_PROP_FRAME_HEIGHT, test_height)
-                            cap.set(cv2.CAP_PROP_FPS, 30)  # Essayer 30 FPS
-                            
-                            # Vérifier si la résolution a été acceptée
-                            actual_width = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
-                            actual_height = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
-                            actual_fps = cap.get(cv2.CAP_PROP_FPS)
-                            
-                            # Tester si on peut lire une frame à cette résolution
-                            ret, frame = cap.read()
-                            if ret and frame is not None and frame.shape[1] >= test_width * 0.9 and frame.shape[0] >= test_height * 0.9:
-                                best_resolution = (actual_width, actual_height)
-                                best_fps = actual_fps
-                                print(f"[CAMERA] Résolution {actual_width}x{actual_height} supportée pour la caméra {i}")
-                                break
-                            else:
-                                print(f"[CAMERA] Résolution {test_width}x{test_height} non supportée pour la caméra {i}")
-                        
-                        if best_resolution:
-                            width, height = best_resolution
-                            fps = best_fps
-                            
-                            # Créer un nom descriptif
-                            backend_name = {
-                                cv2.CAP_ANY: "Auto",
-                                cv2.CAP_DSHOW: "DirectShow",
-                                cv2.CAP_V4L2: "V4L2",
-                                cv2.CAP_GSTREAMER: "GStreamer"
-                            }.get(backend, "Inconnu")
-                            
-                            name = f"Caméra {i} ({backend_name}) - {width}x{height}@{fps:.1f}fps"
-                            available_cameras.append((i, name))
-                            print(f"[CAMERA] ✓ Caméra fonctionnelle détectée: {name}")
-                            break
-                        else:
-                            print(f"[CAMERA] Caméra {i} ouverte mais ne peut pas lire de frame avec backend {backend_name}")
-                    cap.release()
-                except Exception as e:
-                    if cap:
-                        cap.release()
-                    print(f"[CAMERA] Backend {backend} échoué pour caméra {i}: {e}")
-                    continue
-            
-            if not available_cameras or available_cameras[-1][0] != i:
-                print(f"[CAMERA] ✗ Caméra {i} non disponible ou non fonctionnelle")
-                
-        except Exception as e:
-            print(f"[CAMERA] Erreur générale lors de la détection de la caméra {i}: {e}")
-    
-    print(f"[CAMERA] Détection terminée. {len(available_cameras)} caméra(s) fonctionnelle(s) trouvée(s)")
-    return available_cameras
-
-# Classe pour gérer la caméra USB
-class UsbCamera:
-    def __init__(self, camera_id=0):
-        self.camera_id = camera_id
-        self.camera = None
-        self.is_running = False
-        self.thread = None
-        self.frame = None
-        self.lock = threading.Lock()
-        self.error = None
-    
-    def start(self):
-        """Démarrer la caméra USB"""
-        if self.is_running:
-            return True
-        
-        return self._initialize_camera()
-    
-    def _initialize_camera(self):
-        """Initialiser la caméra avec différents backends"""
-        backends = [cv2.CAP_DSHOW, cv2.CAP_ANY, cv2.CAP_V4L2, cv2.CAP_GSTREAMER]
-        
-        for backend in backends:
-            try:
-                backend_name = {
-                    cv2.CAP_ANY: "Auto",
-                    cv2.CAP_DSHOW: "DirectShow",
-                    cv2.CAP_V4L2: "V4L2",
-                    cv2.CAP_GSTREAMER: "GStreamer"
-                }.get(backend, "Inconnu")
-                
-                print(f"[USB CAMERA] Tentative d'ouverture de la caméra {self.camera_id} avec backend {backend_name}...")
-                self.camera = cv2.VideoCapture(self.camera_id, backend)
-                
-                if not self.camera.isOpened():
-                    print(f"[USB CAMERA] Backend {backend_name} : impossible d'ouvrir la caméra {self.camera_id}")
-                    if self.camera:
-                        self.camera.release()
-                    continue
-                
-                # Tester et configurer la meilleure résolution disponible
-                resolutions_to_test = [
-                    (1920, 1080, "Full HD"),  # Full HD
-                    (1280, 720, "HD"),        # HD
-                    (640, 480, "VGA")         # VGA (fallback)
-                ]
-                
-                best_resolution = None
-                for test_width, test_height, res_name in resolutions_to_test:
-                    # Configurer la résolution
-                    self.camera.set(cv2.CAP_PROP_FRAME_WIDTH, test_width)
-                    self.camera.set(cv2.CAP_PROP_FRAME_HEIGHT, test_height)
-                    self.camera.set(cv2.CAP_PROP_FPS, 25)  # Essayer 30 FPS
-                    
-                    # Vérifier la résolution réellement configurée
-                    actual_width = int(self.camera.get(cv2.CAP_PROP_FRAME_WIDTH))
-                    actual_height = int(self.camera.get(cv2.CAP_PROP_FRAME_HEIGHT))
-                    actual_fps = self.camera.get(cv2.CAP_PROP_FPS)
-                    
-                    # Tester si on peut lire une frame à cette résolution
-                    ret, frame = self.camera.read()
-                    if ret and frame is not None and frame.shape[1] >= test_width * 0.9 and frame.shape[0] >= test_height * 0.9:
-                        best_resolution = (actual_width, actual_height, actual_fps, res_name)
-                        print(f"[USB CAMERA] Résolution {res_name} ({actual_width}x{actual_height}@{actual_fps:.1f}fps) configurée avec succès")
-                        break
-                    else:
-                        print(f"[USB CAMERA] Résolution {res_name} ({test_width}x{test_height}) non supportée")
-                
-                if not best_resolution:
-                    print(f"[USB CAMERA] Backend {backend_name} : aucune résolution fonctionnelle trouvée")
-                    self.camera.release()
-                    continue
-                
-                # Vérification finale avec une deuxième lecture
-                ret, frame = self.camera.read()
-                if not ret or frame is None:
-                    print(f"[USB CAMERA] Backend {backend_name} : la caméra {self.camera_id} ne retourne pas d'image de manière stable")
-                    self.camera.release()
-                    continue
-                
-                # Succès !
-                self.is_running = True
-                self.thread = threading.Thread(target=self._capture_loop)
-                self.thread.daemon = True
-                self.thread.start()
-                print(f"[USB CAMERA] Caméra {self.camera_id} démarrée avec succès via backend {backend_name}")
-                return True
-                
-            except Exception as e:
-                print(f"[USB CAMERA] Erreur avec backend {backend_name}: {e}")
-                if self.camera:
-                    self.camera.release()
-                continue
-        
-        # Aucun backend n'a fonctionné
-        self.error = f"Impossible d'ouvrir la caméra {self.camera_id} avec tous les backends testés"
-        print(f"[USB CAMERA] Erreur: {self.error}")
-        return False
-    
-    def _reconnect(self):
-        """Tenter de reconnecter la caméra"""
-        print(f"[USB CAMERA] Tentative de reconnexion de la caméra {self.camera_id}...")
-        if self.camera:
-            self.camera.release()
-        self.camera = None
-        time.sleep(1)  # Attendre un peu avant de réessayer
-        return self._initialize_camera()
-    
-    def _capture_loop(self):
-        """Boucle de capture des frames"""
-        consecutive_errors = 0
-        max_errors = 10  # Nombre maximum d'erreurs consécutives avant de tenter une reconnexion
-        
-        while self.is_running:
-            try:
-                if not self.camera or not self.camera.isOpened():
-                    # Tentative de reconnexion si la caméra est déconnectée
-                    print(f"[USB CAMERA] Caméra {self.camera_id} déconnectée, tentative de reconnexion...")
-                    self._reconnect()
-                    time.sleep(1)  # Attendre avant de réessayer
-                    continue
-                    
-                ret, frame = self.camera.read()
-                if ret:
-                    # Convertir en JPEG pour le streaming MJPEG
-                    _, jpeg = cv2.imencode('.jpg', frame, [cv2.IMWRITE_JPEG_QUALITY, 85])
-                    with self.lock:
-                        self.frame = jpeg.tobytes()
-                    consecutive_errors = 0  # Réinitialiser le compteur d'erreurs
-                else:
-                    consecutive_errors += 1
-                    print(f"[USB CAMERA] Erreur de lecture de frame (tentative {consecutive_errors}/{max_errors})")
-                    if consecutive_errors >= max_errors:
-                        print(f"[USB CAMERA] Trop d'erreurs consécutives, tentative de reconnexion...")
-                        self._reconnect()
-                        consecutive_errors = 0
-                        
-                time.sleep(0.03)  # ~30 FPS
-            except Exception as e:
-                consecutive_errors += 1
-                print(f"[USB CAMERA] Erreur de capture: {e} (tentative {consecutive_errors}/{max_errors})")
-                if consecutive_errors >= max_errors:
-                    print(f"[USB CAMERA] Trop d'erreurs consécutives, tentative de reconnexion...")
-                    self._reconnect()
-                    consecutive_errors = 0
-                time.sleep(0.1)
-    
-    def get_frame(self):
-        """Récupérer la frame actuelle"""
-        with self.lock:
-            return self.frame
-    
-    def stop(self):
-        """Arrêter la caméra"""
-        self.is_running = False
-        if self.thread:
-            self.thread.join(timeout=1.0)
-        if self.camera:
-            self.camera.release()
-        print(f"[USB CAMERA] Caméra {self.camera_id} arrêtée")
 
 # Variables globales
 config = load_config()
@@ -525,20 +172,20 @@ def capture_photo():
                     f.write(last_frame)
                 
                 current_photo = filename
-                print(f"Frame MJPEG capturée avec succès: {filename}")
+                logger.info(f"Frame MJPEG capturée avec succès: {filename}")
                 
                 # Envoyer sur Telegram si activé
                 send_type = config.get('telegram_send_type', 'photos')
                 if send_type in ['photos', 'both']:
-                    threading.Thread(target=send_to_telegram, args=(filepath, "photo")).start()
+                    threading.Thread(target=send_to_telegram, args=(filepath, config, "photo")).start()
                 
                 return jsonify({'success': True, 'filename': filename})
             else:
-                print("Aucune frame disponible dans le flux")
+                logger.info("Aucune frame disponible dans le flux")
                 return jsonify({'success': False, 'error': 'Aucune frame disponible'})
             
     except Exception as e:
-        print(f"Erreur lors de la capture: {e}")
+        logger.info(f"Erreur lors de la capture: {e}")
         return jsonify({'success': False, 'error': f'Erreur de capture: {str(e)}'})
 
 @app.route('/review')
@@ -662,7 +309,7 @@ def apply_effect():
         return result
             
     except Exception as e:
-        print(f"Erreur lors de l'application de l'effet: {e}")
+        logger.info(f"Erreur lors de l'application de l'effet: {e}")
         return jsonify({'success': False, 'error': f'Erreur IA: {str(e)}'})
 
 async def apply_effect_async(photo_path):
@@ -670,27 +317,27 @@ async def apply_effect_async(photo_path):
     global current_photo
     
     try:
-        print("[DEBUG IA] Début de l'application de l'effet IA")
-        print(f"[DEBUG IA] Photo source: {photo_path}")
-        print(f"[DEBUG IA] Clé API configurée: {'Oui' if config.get('runware_api_key') else 'Non'}")
-        print(f"[DEBUG IA] Prompt: {config.get('effect_prompt', 'Transform this photo into a beautiful ghibli style')}")
+        logger.info("[DEBUG IA] Début de l'application de l'effet IA")
+        logger.info(f"[DEBUG IA] Photo source: {photo_path}")
+        logger.info(f"[DEBUG IA] Clé API configurée: {'Oui' if config.get('runware_api_key') else 'Non'}")
+        logger.info(f"[DEBUG IA] Prompt: {config.get('effect_prompt', 'Transform this photo into a beautiful ghibli style')}")
         
         # Initialiser Runware
-        print("[DEBUG IA] Initialisation de Runware...")
+        logger.info("[DEBUG IA] Initialisation de Runware...")
         runware = Runware(api_key=config['runware_api_key'])
-        print("[DEBUG IA] Connexion à Runware...")
+        logger.info("[DEBUG IA] Connexion à Runware...")
         await runware.connect()
-        print("[DEBUG IA] Connexion établie avec succès")
+        logger.info("[DEBUG IA] Connexion établie avec succès")
         
         # Lire et encoder l'image en base64
-        print("[DEBUG IA] Lecture et encodage de l'image...")
+        logger.info("[DEBUG IA] Lecture et encodage de l'image...")
         with open(photo_path, 'rb') as img_file:
             img_data = img_file.read()
             img_base64 = base64.b64encode(img_data).decode('utf-8')
-        print(f"[DEBUG IA] Image encodée: {len(img_base64)} caractères base64")
+        logger.info(f"[DEBUG IA] Image encodée: {len(img_base64)} caractères base64")
         
         # Préparer la requête d'inférence avec referenceImages (requis pour ce modèle)
-        print("[DEBUG IA] Préparation de la requête d'inférence avec referenceImages...")
+        logger.info("[DEBUG IA] Préparation de la requête d'inférence avec referenceImages...")
         request = IImageInference(
             positivePrompt=config.get('effect_prompt', 'Transforme cette image en illustration de style Studio Ghibli'),
             referenceImages=[f"data:image/jpeg;base64,{img_base64}"],
@@ -701,55 +348,55 @@ async def apply_effect_async(photo_path):
             CFGScale=2.5,
             numberResults=1
         )
-        print("[DEBUG IA] Requête préparée avec les paramètres de base:")
-        print(f"[DEBUG IA]   - Modèle: runware:106@1")
-        print(f"[DEBUG IA]   - Dimensions: 1392x752")
-        print(f"[DEBUG IA]   - Étapes: {config.get('effect_steps', 5)}")
-        print(f"[DEBUG IA]   - CFG Scale: 2.5")
-        print(f"[DEBUG IA]   - Nombre de résultats: 1")
+        logger.info("[DEBUG IA] Requête préparée avec les paramètres de base:")
+        logger.info(f"[DEBUG IA]   - Modèle: runware:106@1")
+        logger.info(f"[DEBUG IA]   - Dimensions: 1392x752")
+        logger.info(f"[DEBUG IA]   - Étapes: {config.get('effect_steps', 5)}")
+        logger.info(f"[DEBUG IA]   - CFG Scale: 2.5")
+        logger.info(f"[DEBUG IA]   - Nombre de résultats: 1")
         
         # Appliquer l'effet
-        print("[DEBUG IA] Envoi de la requête à l'API Runware...")
+        logger.info("[DEBUG IA] Envoi de la requête à l'API Runware...")
         # La méthode correcte est imageInference
         images = await runware.imageInference(requestImage=request)
-        print(f"[DEBUG IA] Réponse reçue: {len(images) if images else 0} image(s) générée(s)")
+        logger.info(f"[DEBUG IA] Réponse reçue: {len(images) if images else 0} image(s) générée(s)")
         
         if images and len(images) > 0:
             # Télécharger l'image transformée
-            print(f"[DEBUG IA] URL de l'image générée: {images[0].imageURL}")
-            print("[DEBUG IA] Téléchargement de l'image transformée...")
+            logger.info(f"[DEBUG IA] URL de l'image générée: {images[0].imageURL}")
+            logger.info("[DEBUG IA] Téléchargement de l'image transformée...")
             import requests
             response = requests.get(images[0].imageURL)
-            print(f"[DEBUG IA] Statut de téléchargement: {response.status_code}")
+            logger.info(f"[DEBUG IA] Statut de téléchargement: {response.status_code}")
             
             if response.status_code == 200:
-                print(f"[DEBUG IA] Taille de l'image téléchargée: {len(response.content)} bytes")
+                logger.info(f"[DEBUG IA] Taille de l'image téléchargée: {len(response.content)} bytes")
                 
                 # S'assurer que le dossier effet existe
-                print(f"[DEBUG IA] Vérification du dossier effet: {EFFECT_FOLDER}")
+                logger.info(f"[DEBUG IA] Vérification du dossier effet: {EFFECT_FOLDER}")
                 os.makedirs(EFFECT_FOLDER, exist_ok=True)
-                print(f"[DEBUG IA] Dossier effet existe: {os.path.exists(EFFECT_FOLDER)}")
+                logger.info(f"[DEBUG IA] Dossier effet existe: {os.path.exists(EFFECT_FOLDER)}")
                 
                 # Créer un nouveau nom de fichier pour l'image avec effet
                 timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
                 effect_filename = f'effect_{timestamp}.jpg'
                 effect_path = os.path.join(EFFECT_FOLDER, effect_filename)
-                print(f"[DEBUG IA] Sauvegarde vers: {effect_path}")
+                logger.info(f"[DEBUG IA] Sauvegarde vers: {effect_path}")
                 
                 # Sauvegarder l'image avec effet
                 with open(effect_path, 'wb') as f:
                     f.write(response.content)
-                print("[DEBUG IA] Image sauvegardée avec succès")
+                logger.info("[DEBUG IA] Image sauvegardée avec succès")
                 
                 # Mettre à jour la photo actuelle
                 current_photo = effect_filename
-                print(f"[DEBUG IA] Photo actuelle mise à jour: {current_photo}")
-                print("[DEBUG IA] Effet appliqué avec succès!")
+                logger.info(f"[DEBUG IA] Photo actuelle mise à jour: {current_photo}")
+                logger.info("[DEBUG IA] Effet appliqué avec succès!")
                 
                 # Envoyer sur Telegram si activé
                 send_type = config.get('telegram_send_type', 'photos')
                 if send_type in ['effet', 'both']:
-                    threading.Thread(target=send_to_telegram, args=(effect_path, "effet")).start()
+                    threading.Thread(target=send_to_telegram, args=(effect_path, config, "effet")).start()
                 
                 return jsonify({
                     'success': True, 
@@ -757,14 +404,14 @@ async def apply_effect_async(photo_path):
                     'new_filename': effect_filename
                 })
             else:
-                print(f"[DEBUG IA] ERREUR: Échec du téléchargement (code {response.status_code})")
+                logger.info(f"[DEBUG IA] ERREUR: Échec du téléchargement (code {response.status_code})")
                 return jsonify({'success': False, 'error': 'Erreur lors du téléchargement de l\'image transformée'})
         else:
-            print("[DEBUG IA] ERREUR: Aucune image générée par l'IA")
+            logger.info("[DEBUG IA] ERREUR: Aucune image générée par l'IA")
             return jsonify({'success': False, 'error': 'Aucune image générée par l\'IA'})
             
     except Exception as e:
-        print(f"Erreur lors de l'application de l'effet: {e}")
+        logger.info(f"Erreur lors de l'application de l'effet: {e}")
         return jsonify({'success': False, 'error': f'Erreur IA: {str(e)}'})
 
 @app.route('/admin')
@@ -881,7 +528,7 @@ def save_admin_config():
         
         # Configuration de l'imprimante
         config['printer_enabled'] = 'printer_enabled' in request.form
-        config['printer_port'] = request.form.get('printer_port', '/dev/AMA0')
+        config['printer_port'] = request.form.get('printer_port', '/dev/ttyAMA0')
         
         printer_baudrate = request.form.get('printer_baudrate', '9600').strip()
         try:
@@ -1055,7 +702,7 @@ def generate_video_stream():
         
         # Utiliser la caméra USB si configurée
         if camera_type == 'usb':
-            print("[CAMERA] Démarrage de la caméra USB...")
+            logger.info("[CAMERA] Démarrage de la caméra USB...")
             camera_id = config.get('usb_camera_id', 0)
             usb_camera = UsbCamera(camera_id=camera_id)
             if not usb_camera.start():
@@ -1079,7 +726,7 @@ def generate_video_stream():
         
         # Utiliser la Pi Camera par défaut
         else:
-            print("[CAMERA] Démarrage de la Pi Camera...")
+            logger.info("[CAMERA] Démarrage de la Pi Camera...")
             # Commande libcamera-vid pour flux MJPEG - résolution 16/9
             cmd = [
                 'libcamera-vid',
@@ -1140,11 +787,11 @@ def generate_video_stream():
                                jpeg_frame + b'\r\n')
                                
                 except Exception as e:
-                    print(f"[CAMERA] Erreur lecture flux: {e}")
+                    logger.info(f"[CAMERA] Erreur lecture flux: {e}")
                     break
                 
     except Exception as e:
-        print(f"Erreur flux vidéo: {e}")
+        logger.info(f"Erreur flux vidéo: {e}")
         # Envoyer une frame d'erreur
         error_msg = f"Erreur caméra: {str(e)}"
         yield (b'--frame\r\n'
@@ -1162,7 +809,7 @@ def stop_camera_process():
         try:
             usb_camera.stop()
         except Exception as e:
-            print(f"[CAMERA] Erreur lors de l'arrêt de la caméra USB: {e}")
+            logger.info(f"[CAMERA] Erreur lors de l'arrêt de la caméra USB: {e}")
         usb_camera = None
     
     # Arrêter le processus libcamera-vid si actif
@@ -1195,7 +842,7 @@ def stop_camera():
 # Nettoyer les processus à la fermeture
 @atexit.register
 def cleanup():
-    print("[APP] Arrêt de l'application, nettoyage des ressources...")
+    logger.info("[APP] Arrêt de l'application, nettoyage des ressources...")
     stop_camera_process()
 
 def signal_handler(sig, frame):
